requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.18.3
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
arxiv==2.2.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
catboost==1.2.8
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
databricks-sdk==0.85.0
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
# Ember dependency removed for remote stability
# -e d:\disk d\study material\ai & ml\ai_malw_analysis\dataset\ember-master
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.12
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.9.23
fonttools==4.55.3
frozenlist==1.7.0
fsspec==2025.12.0
gast==0.6.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.48.0
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huey==2.6.0
hyperlink==21.0.0
idna==3.10
ImageIO==2.37.2
importlib_metadata==8.7.1
incremental==24.7.2
ipykernel==6.29.5
ipython==8.31.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.12.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.12.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.26
lazy_loader==0.4
libclang==18.1.1
lief==0.17.1
lightgbm==4.6.0
llvmlite==0.46.0
lxml==5.3.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mlflow==3.9.0
mlflow-skinny==3.9.0
mlflow-tracing==3.9.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.22.0
nest-asyncio==1.6.0
networkx==3.6.1
node==1.2.2
numba==0.63.1
numpy==2.0.1
odict==1.9.0
openai==2.8.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==24.2
pandas==2.2.2
parsel==1.10.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.5.0
plumber==1.7
plyer==2.1.0
prettytable==3.17.0
prompt_toolkit==3.0.48
propcache==0.3.2
Protego==0.4.0
protobuf==5.29.6
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.18.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==6.0.0
pyshark==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.7.0
referencing==0.35.1
regex==2025.11.3
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
scapy==2.6.1
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.14.1
Scrapy==2.12.0
seaborn==0.13.2
sed==0.3.1
service-identity==24.2.0
setuptools==75.1.0
sgmllib3k==1.0.0
shap==0.50.0
six==1.16.0
skops==0.13.0
slicer==0.0.8
smart_open==7.4.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
source==1.2.0
SQLAlchemy==2.0.43
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.46.2
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.5.0
tifffile==2025.12.12
tiktoken==0.12.0
tldextract==5.1.3
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.34.2
w3lib==2.3.1
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
xgboost==3.1.1
yarl==1.20.1
zipp==3.23.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==5.3
zstandard==0.24.0

# --- PolicyCortex Missing Dependencies ---
reportlab==4.3.1
llama-cpp-python==0.3.7
sentence-transformers==3.4.1
faiss-cpu==1.10.0
google-generativeai==0.8.3
